ODBIdentityDictionary defines a very odd method not found in any other dictionary in Pharo:  #at:ifAbsentPutValue:
This is the same as at:ifAbsentPut:, but not sending #value to the block/object to to be added. This allows the user to not wrap the argument
in a block, which is slow as the block is created at runtime.

In Pharo this is not needed, as we have implemented #value on Object. This allwos us to use the "no block" pattern with at:ifAbsentPut:

This PR 
	- changes the one sender 
	-  ODBIdentityDictionary should implement  at:ifAbsentPut: